Added support for unrecognized entities to the RoughTok stage
In the former specification outlined by Ondrej Bojar, unrecognized
entities in the text were to be treated as a single token. I have
modified the quex file for rough tokenization (which sadly resulted
in doubling the number of modes) so it now cuts out entities as
whole solitary tokens at no expense of robustness (all edge cases
still perform well).