Explicitly adding negative features
Slighlty helps to prevent the maxent library from getting into
the situation when it encounters a previously unseen predicate.
Experiments and examination of the maxent library could lead us
to some estimates as to how much does that affect the accuracy
of the model (I don't think it should, but on the toy case I was
testing it seemed to slightly improve the accuracy). The reason
why this change isn't in master is because it seems to be a real
performance drain on maxent.